requirements.txt
PyYAML==5.1
selenium==3.141.0
urllib3==1.24.1
